Replace old scripts with new data and R(2+1)D structure
